requirements_tg.txt
numpy==1.26.4
scikit-learn==1.3.2
lightgbm==4.5.0
aiogram==2.15
python-dotenv>=0.19.0
joblib>=1.2.0
clickhouse-driver>=0.2.0
nltk>=3.8.1